backend/azure_functions/requirements.txt
azure-functions
azure-cosmos==4.5.1
requests
python-dotenv
pandas
